SmartCar_backend/SmartCart/requirements.txt
Django>=5.1.2
Pillow>=10.0.0
fal-client>=0.5.0
requests>=2.31.0
python-dotenv>=1.0.0
opencv-python>=4.8.0
numpy>=1.24.0
stripe>=7.0.0
